+++ src/worker/task_pca.py
@@ -1,7 +1,5 @@
 # stdlib
-import os
 import pickle
-import time
 import logging
 
 # third party lib
@@ -16,6 +14,7 @@ from lib.corr import nancorr
 from lib.utils import write_or_replace
 
 logger = logging.getLogger("worker")
+
 
 class LdReporter(object):
     __instance = None
@@ -44,7 +43,7 @@ class LdReporter(object):
         data = pickle.loads(data)
         n = self.store.attrs['n']
         msg = {}
-        if self.r3 == 0: # fake start the first round
+        if self.r3 == 0:  # fake start the first round
             for chrom in self.chroms:
                 # if the length is less than r1, you deserve an error.
                 # No apologies
@@ -55,8 +54,8 @@ class LdReporter(object):
                 continue
             chrom = key
             tags = self.store["{}/PCA_passed".format(chrom)]
-            if state[0] == "E": # Finished with this chrom
-                if len(data) == 1: # Done with everything
+            if state[0] == "E":  # Finished with this chrom
+                if len(data) == 1:  # Done with everything
                     msg = pickle.dumps({})
                     networking.respond_to_server('api/tasks/PCA/PCAPOS', 'POST', msg,
                                                  client_config['name'], env)
@@ -70,9 +69,9 @@ class LdReporter(object):
             pos = self.store["{}/PCA_positions".format(chrom)]
             positions = pos[self.r3: end]
             positions = positions[tokeep]
-            genotypes = np.empty((n,len(positions)), dtype=np.float32)
+            genotypes = np.empty((n, len(positions)), dtype=np.float32)
             for i, snp in enumerate(positions):
-                genotypes[:,i] = self.store["{}/{}".format(chrom, snp)].value
+                genotypes[:, i] = self.store["{}/{}".format(chrom, snp)].value
             corr = nancorr(genotypes)
             msg[chrom] = corr
         msg = pickle.dumps(msg)
@@ -91,7 +90,7 @@ def store_filtered(message, client_config):
             mask = store["{}/PCA_mask".format(chrom)]
             if "non_ld_mask" not in store[chrom]:
                 pcmask = mask.value
-                dset = store.require_dataset(f"{chrom}/non_ld_mask", pcmask.shape, dtype = pcmask.dtype)
+                dset = store.require_dataset(f"{chrom}/non_ld_mask", pcmask.shape, dtype=pcmask.dtype)
                 dset[:] = pcmask
             else:
                 pcmask = store[f"{chrom}/non_ld_mask"].value
@@ -103,8 +102,8 @@ def report_cov(client_config, env):
     def standardize_mat(mat, af, sd):
         af = 2 * af.reshape(af.shape[0], 1)
         mat -= af
-        ind = sd>0
-        mat[ind, :] /= sd[ind].reshape(np.sum(ind),1)
+        ind = sd > 0
+        mat[ind, :] /= sd[ind].reshape(np.sum(ind), 1)
         mat[np.isnan(mat)] = 0
         return mat
     pfile = shared.get_plink_store(client_config["plinkfile"])
@@ -131,8 +130,8 @@ def report_cov(client_config, env):
                 tokeep = group['PCA_mask'].value
                 af2 = group["MAF"].value[tokeep]
                 sd2 = np.sqrt(group["VAR"].value[tokeep])
-                pos   = group["positions"].value[tokeep]
-                g2     = np.empty((n, len(pos)))
+                pos = group["positions"].value[tokeep]
+                g2 = np.empty((n, len(pos)))
                 for i, snp2 in enumerate(pos):
                     g2[:, i] = group[str(snp2)].value
                 g2 = standardize_mat(g2.transpose(), af2, sd2).transpose()
@@ -143,7 +142,7 @@ def report_cov(client_config, env):
                 if ch1 == chroms[-1] and ch2 == chroms[-1]:
                     msg["E"] = True
                 msg = pickle.dumps(msg)
-                #time.sleep(1)
+                # time.sleep(1)
                 networking.respond_to_server('api/tasks/PCA/COV', 'POST', msg, client_config['name'], env)
         logger.info(f"Final size will be {size}")
 
@@ -159,13 +158,13 @@ def pca_projection(data, client_config):
         for chrom in chroms:
             n += np.sum(store[f"{chrom}/PCA_mask"])
         num_inds = store.attrs["n"]
-        #pca_sigma = dset.require_dataset('pca_sigma', shape=inv_sigma.shape, dtype=np.float32)
-        #pca_sigma[:] = inv_sigma
+        # pca_sigma = dset.require_dataset('pca_sigma', shape=inv_sigma.shape, dtype=np.float32)
+        # pca_sigma[:] = inv_sigma
         arr = np.empty((num_inds, n), dtype=np.float32)
         offset = 0
         for chrom in chroms:
             group = store[str(chrom)]
-            tokeep    = group["PCA_mask"].value
+            tokeep = group["PCA_mask"].value
             af = group["MAF"].value[tokeep]
             sd = np.sqrt(group["VAR"].value[tokeep])
             positions = group["positions"].value[tokeep]
@@ -180,10 +179,10 @@ def pca_projection(data, client_config):
         write_or_replace(dset, 'pca_sigma', val=inv_sigma)
         write_or_replace(dset, 'pca_v.T', val=v)
         write_or_replace(dset, 'pca_u', val=u)
-        #pca_vt = dset.require_dataset('pca_v.T', shape=v.shape,
+        # pca_vt = dset.require_dataset('pca_v.T', shape=v.shape,
         #    dtype=np.float32)
-        #pca_vt[:,:] = v
-        #pca_u = dset.require_dataset('pca_u', shape=u.shape,
+        # pca_vt[:,:] = v
+        # pca_u = dset.require_dataset('pca_u', shape=u.shape,
         #    dtype=np.float32)
-        #pca_u[:,:] = u
+        # pca_u[:,:] = u
     logger.info("Done with projection!")
